Refactor: Move FavButton to its own component
