+++ src/components/Session.vue
@@ -22,9 +22,7 @@ a.c-linear-schedule-session(:class="{faved}", :style="style", :href="link", @cli
 			.track(v-if="session.track") {{ getLocalizedString(session.track.name) }}
 			.room(v-if="showRoom && session.room") {{ getLocalizedString(session.room.name) }}
 	.session-icons
-		bunt-icon-button.btn-fav-container(@click.prevent.stop="toggleFav")
-			svg.star(viewBox="0 0 24 24", ref="star")
-				path(d="M12,17.27L18.18,21L16.54,13.97L22,9.24L14.81,8.62L12,2L9.19,8.62L2,9.24L7.45,13.97L5.82,21L12,17.27Z")
+		fav-button(@toggleFav="toggleFav")
 		svg.do-not-record(v-if="session.do_not_record", viewBox="0 0 116.59076 116.59076", width="4116.59076mm", height="116.59076mm", fill="none", xmlns="http://www.w3.org/2000/svg")
 			g(transform="translate(-9.3465481,-5.441411)")
 				rect(style="fill:#000000;fill-opacity;stroke:none;stroke-width:11.2589;stroke-linecap:round;stroke-dasharray:none;stroke-opacity:1;paint-order:markers stroke fill", width="52.753284", height="39.619537", x="35.496307", y="43.927021", rx="5.5179553", ry="7.573648")
@@ -36,6 +34,7 @@ a.c-linear-schedule-session(:class="{faved}", :style="style", :href="link", @cli
 import { DateTime } from 'luxon'
 import MarkdownIt from 'markdown-it'
 import { getLocalizedString, getPrettyDuration, getSessionTime } from '~/utils'
+import FavButton from '~/components/FavButton.vue'
 
 const markdownIt = MarkdownIt({
 	linkify: true,
@@ -87,6 +86,9 @@ export default {
 			}
 		}
 	},
+	components: {
+		FavButton
+	},
 	data () {
 		return {
 			getPrettyDuration,
@@ -125,15 +127,12 @@ export default {
 	},
 	methods: {
 		toggleFav () {
+			console.log("toggling fav")
 			if (this.faved) {
 				this.$emit('unfav', this.session.id)
 			} else {
 				this.$emit('fav', this.session.id)
 			}
-			this.$refs.star.classList.add('rotate-star')
-			setTimeout(() => {
-				this.$refs.star.classList.remove('rotate-star')
-			}, 400)
 		}
 	}
 }
@@ -256,21 +255,9 @@ export default {
 			padding: 2px
 			width: 32px
 			height: 32px
-			svg.star
-				height: 20px
-				width: 20px
-				path
-					fill: none
-					stroke: #ffa000
-					stroke-width: 1px
-					vector-effect: non-scaling-stroke
 	&.faved
 		.btn-fav-container
 			display: inline-flex
-			svg.star
-				filter: drop-shadow(0 0 2px rgba(0 0 0 0.17))
-				path
-					fill: #ffa000
 	&:hover
 		.info
 			border: 1px solid var(--track-color)
@@ -279,13 +266,4 @@ export default {
 				color: var(--pretalx-clr-primary)
 		.btn-fav-container
 			display: inline-flex
-	// +below('m')
-	// 	min-width: 0
-	svg.star.rotate-star
-		animation: rotate 0.4s
-	@keyframes rotate
-		0%
-			transform: rotate(0deg)
-		100%
-			transform: rotate(72deg)
 </style>
